+++ sdk/src/main/java/com/bugsnag/android/DefaultDelivery.java
@@ -2,6 +2,7 @@ package com.bugsnag.android;
 
 import android.net.ConnectivityManager;
 import android.net.NetworkInfo;
+import android.support.annotation.Nullable;
 
 import java.io.BufferedWriter;
 import java.io.IOException;
@@ -14,9 +15,10 @@ import java.util.Map;
 
 class DefaultDelivery implements Delivery {
 
-    private final ConnectivityManager connectivityManager;
+    private static final int HTTP_REQUEST_FAILED = 0;
+    @Nullable private final ConnectivityManager connectivityManager;
 
-    DefaultDelivery(ConnectivityManager connectivityManager) {
+    DefaultDelivery(@Nullable ConnectivityManager connectivityManager) {
         this.connectivityManager = connectivityManager;
     }
 
@@ -79,12 +81,19 @@ class DefaultDelivery implements Delivery {
             return conn.getResponseCode();
         } catch (IOException exception) {
             throw new DeliveryFailureException("IOException encountered in request", exception);
+        } catch (Exception exception) {
+            Logger.warn("Unexpected error delivering payload", exception);
+            return HTTP_REQUEST_FAILED;
         } finally {
             IOUtils.close(conn);
         }
     }
 
     private void checkHasNetworkConnection() throws DeliveryFailureException {
+        if (connectivityManager == null) {
+            return; // unlikely case, allow delivery attempt without checking connection first
+        }
+
         NetworkInfo activeNetworkInfo = connectivityManager.getActiveNetworkInfo();
 
         // conserve device battery by avoiding radio use